FLUID-5695: Fixes following review - also, added npm scripts for easy running of docpad workflow
