+++ package.json
@@ -1,15 +1,19 @@
 {
-  "name": "infusion-docs",
-  "version": "1.5.0",
-  "description": "Infusion Documentation",
-  "license" : "CC-BY-3.0",
-  "repository": "https://github.com/fluid-project/infusion-docs",
-  "dependencies": {
-    "URIjs": "^1.14.1",
-    "docpad-plugin-ghpages": "~2.4.4",
-    "docpad-plugin-handlebars": "~2.3.0",
-    "docpad-plugin-highlightjs": "~2.2.2",
-    "docpad-plugin-marked": "~2.2.1",
-    "fs-extra": "^0.18.0"
-  }
+    "name": "infusion-docs",
+    "version": "1.5.0",
+    "description": "Infusion Documentation",
+    "license" : "CC-BY-3.0",
+    "repository": "https://github.com/fluid-project/infusion-docs",
+    "scripts": {
+        "docpad": "docpad run -c docpad-config.js --env static",
+        "deploy": "docpad deploy-ghpages -c docpad-config.js --env static"
+    },
+    "dependencies": {
+        "URIjs": "^1.14.1",
+        "docpad-plugin-ghpages": "~2.4.4",
+        "docpad-plugin-handlebars": "~2.3.0",
+        "docpad-plugin-highlightjs": "~2.2.2",
+        "docpad-plugin-marked": "~2.2.1",
+        "fs-extra": "^0.18.0"
+    }
 }
