+++ src/documents/ChangeApplierAPI.md
@@ -412,7 +412,7 @@ Users will in most cases only be interested in the first argument in this signat
 
 ### Declarative style for triggering a change###
 
-The declarative style for triggering model changes involve a kind of IoC record (a _change record_) that is supported in various places in component configuration, 
+The declarative style for triggering model changes involves an IoC record (a _change record_) that is supported in various places in component configuration, 
 in particular as part of the definition of both [Invokers](Invokers.md) and [Listeners](InfusionEventSystem.md) of an IoC-configured component.
 This style of record is recognised by its use of the special member `changePath` (a "[duck typing](https://en.wikipedia.org/wiki/Duck_typing) field") which determines which path in which component model will receive the change.
 
@@ -534,8 +534,9 @@ applier.fireChangeRequest(<changeRequest>)
 where a `changeRequest` is an object holding the above named parameters in named fields - e.g. `{path: "modelPath", value: "newValue"}`. Note that a `changeRequest` is the same as a 
 [`changeRecord`](#declarative-style-for-triggering-a-change) only the path is encoded in a field named `path` rather than `changePath`.
 
-`change` and `fireChangeRequest` reach exactly the same implementation - the only difference is in the packaging of the arguments. For `change` they are spread out as a sequence of 4 arguments, whereas for `fireChangeRequest`, they are packaged up as named fields (`path`, `value` and `type`) of a plain JavaScript object. 
-Such an object is called a **changeRequest** and is a convenient package for these requests to pass around in in an event pipeline within the framework.
+`change` and `fireChangeRequest` reach exactly the same implementation - the only difference is in the packaging of the arguments. For `change` they are spread out as a sequence of 4 arguments, 
+whereas for `fireChangeRequest`, they are packaged up as named fields (`path`, `value` and `type`) of a plain JavaScript object. 
+Such an object is called a **changeRequest** and is a convenient package for these requests to pass around in an event pipeline within the framework.
 
 The programmatic style for **firing** changes is less strongly discouraged than the programmatic style for **listening** to changes is - since it does not run into the same lifecycle issues that programmatic listeners do. However, the declarative style for triggering changes should be used wherever it can.
